+++ Twig/Extension/AbstractLocaleAwareExtension.php
@@ -2,7 +2,6 @@
 
 namespace Craue\TwigExtensionsBundle\Twig\Extension;
 
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Twig\Extension\AbstractExtension;
 
@@ -21,17 +20,12 @@ abstract class AbstractLocaleAwareExtension extends AbstractExtension {
 	protected $locale = 'en-US';
 
 	/**
-	 * @var ContainerInterface|null
-	 */
-	protected $container;
-
-	/**
 	 * @var RequestStack|null
 	 */
 	private $requestStack;
 
 	/**
-	 * @param mixed $value The request stack, the service container or a locale string.
+	 * @param mixed $value The request stack or a locale string.
 	 * @throws \InvalidArgumentException
 	 */
 	public function setLocale($value) {
@@ -49,16 +43,9 @@ abstract class AbstractLocaleAwareExtension extends AbstractExtension {
 			return;
 		}
 
-		if ($value instanceof ContainerInterface) {
-			@trigger_error(sprintf('Passing the service container to "%s" is deprecated. Instead, pass either the request stack or a locale string.', __METHOD__), E_USER_DEPRECATED);
-			$this->container = $value;
-			return;
-		}
-
 		throw new \InvalidArgumentException(sprintf(
-				'Expected argument of either type "string", "%s", or "%s", but "%s" given.',
+				'Expected argument of either type "string" or "%s", but "%s" given.',
 				RequestStack::class,
-				ContainerInterface::class,
 				is_object($value) ? get_class($value) : gettype($value)
 		));
 	}
@@ -67,10 +54,6 @@ abstract class AbstractLocaleAwareExtension extends AbstractExtension {
 	 * @return string
 	 */
 	public function getLocale() {
-		if ($this->container !== null) {
-			$this->requestStack = $this->container->get('request_stack');
-		}
-
 		if ($this->requestStack !== null) {
 			$currentRequest = $this->requestStack->getCurrentRequest();
 			if ($currentRequest !== null) {